Pr4 - Add employee info

+++ src/main/java/com/example/nom035/repository/OptionAnswerRepository.java
@@ -4,6 +4,6 @@ import com.example.nom035.entity.OptionAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface OptionAnswerRepository extends JpaRepository<OptionAnswer, Integer> {
-    List<OptionAnswer> findByQuestionId(Integer questionId);
+public interface OptionAnswerRepository extends JpaRepository<OptionAnswer, Long> {
+    List<OptionAnswer> findByQuestionId(Long questionId);
 }
