+++ src/main/java/com/example/nom035/repository/QuestionRepository.java
@@ -4,6 +4,6 @@ import com.example.nom035.entity.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
-    List<Question> findBySurveyId(Integer surveyId);
+public interface QuestionRepository extends JpaRepository<Question, Long> {
+    List<Question> findBySurveyId(Long surveyId);
 }
