+++ src/main/java/com/example/nom035/repository/ResponseRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ResponseRepository extends JpaRepository<Response, Integer> {
-    List<Response> findBySurveyApplicationId(Integer surveyApplicationId);
-    List<Response> findByQuestionId(Integer questionId);   
+public interface ResponseRepository extends JpaRepository<Response, Long> {
+    List<Response> findBySurveyApplicationId(Long surveyApplicationId);
+    List<Response> findByQuestionId(Long questionId);   
     @Query("SELECT r FROM Response r WHERE r.surveyApplication.employee.company.id = :companyId")
     List<Response> findAllByCompanyId(@Param("companyId") Long companyId);
 }
