+++ src/main/java/com/example/nom035/repository/SurveyApplicationRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface SurveyApplicationRepository extends JpaRepository<SurveyApplication, Integer> {
-    List<SurveyApplication> findByCompanySurveyId(Integer companySurveyId);
-    List<SurveyApplication> findByEmployeeId(Integer employeeId);
+public interface SurveyApplicationRepository extends JpaRepository<SurveyApplication, Long> {
+    List<SurveyApplication> findByCompanySurveyId(Long companySurveyId);
+    List<SurveyApplication> findByEmployeeId(Long employeeId);
     @Query("SELECT sa.status, COUNT(sa) FROM SurveyApplication sa " +
             "WHERE sa.employee.company.id = :companyId GROUP BY sa.status")
      List<Object[]> countByStatusAndCompanyId(@Param("companyId") Long companyId);
